+++ backEnd/TT.SoMall.sln
@@ -45,17 +45,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Abp.Shops", "modules\TT.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Abp.OssManagement", "modules\TT.Abp.OssManagement\TT.Abp.OssManagement.csproj", "{7CFCDAA2-B321-4260-A603-1268E3FEF27A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Core", "modules\TT.Core\TT.Core.csproj", "{B6DCCB30-EB9D-476E-9D7D-A57782C2B1CB}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Lib", "Lib", "{3822D417-4A42-436D-B81E-8CF5517BE161}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Extensions", "modules\TT.Extensions\TT.Extensions.csproj", "{B6DCCB30-EB9D-476E-9D7D-A57782C2B1CB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Abp.FormManagement", "modules\TT.Abp.FormManagement\TT.Abp.FormManagement.csproj", "{7A73ED08-BA3C-4E6B-8DD6-972E5D80A48E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Abp.VisitorManagement", "modules\TT.Abp.VisitorManagement\TT.Abp.VisitorManagement.csproj", "{EE8638FC-6C61-43EC-B335-5A5368328751}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.Abp.Weixin", "modules\TT.Abp.Weixin\TT.Abp.Weixin.csproj", "{94F0E29F-6762-4244-AB42-BC8BBD30F3B4}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "httpClient_modules", "httpClient_modules", "{BD097154-4B9C-4867-B154-4739E8A96DBB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.HttpClient.Weixin", "modules\TT.HttpClient.Weixin\TT.HttpClient.Weixin.csproj", "{6B4D1543-D8B7-4551-AE50-CA048D18AB4C}"
 EndProject
@@ -248,11 +244,9 @@ Global
 		{AA94D832-1CCC-4715-95A9-A483F23A1A5D} = {CA9AC87F-097E-4F15-8393-4BC07735A5B0}
 		{D9204363-35E8-4400-BFC4-4700105AFBC7} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{7CFCDAA2-B321-4260-A603-1268E3FEF27A} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
-		{B6DCCB30-EB9D-476E-9D7D-A57782C2B1CB} = {3822D417-4A42-436D-B81E-8CF5517BE161}
 		{7A73ED08-BA3C-4E6B-8DD6-972E5D80A48E} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{EE8638FC-6C61-43EC-B335-5A5368328751} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{94F0E29F-6762-4244-AB42-BC8BBD30F3B4} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
-		{6B4D1543-D8B7-4551-AE50-CA048D18AB4C} = {BD097154-4B9C-4867-B154-4739E8A96DBB}
 		{D0A4CF65-2AA3-484A-B853-ADC37999B66F} = {44EEB9F6-B409-470E-A95A-E5F5F6CDEDC6}
 		{8891565C-1100-4B77-BEFF-16F58B3ABD88} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{B266681B-5A27-43BB-9BD3-CC10993C2D8A} = {04DBDB01-70F4-4E06-B468-8F87850B22BE}
@@ -261,8 +255,10 @@ Global
 		{E8BDFF2D-CCD0-40E5-A791-2283E355D794} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{199423D3-56F2-4BFF-805F-A2C193F8C916} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 		{9C67424B-30A8-4A21-815F-8CCC56067E72} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
-		{B7CF64FB-5583-4D0B-AE32-E630719A1458} = {3822D417-4A42-436D-B81E-8CF5517BE161}
 		{84E22391-632C-494D-8B13-C9928A264C22} = {BA342A16-F659-4216-BA72-635ABF469A9F}
+		{B7CF64FB-5583-4D0B-AE32-E630719A1458} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
+		{B6DCCB30-EB9D-476E-9D7D-A57782C2B1CB} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
+		{6B4D1543-D8B7-4551-AE50-CA048D18AB4C} = {6B55CB5D-72E2-4AB1-844C-227E2C2934DA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {28315BFD-90E7-4E14-A2EA-F3D23AF4126F}
