+++ backEnd/modules/TT.Abp.Audit/TT.Abp.AuditManagement.csproj
@@ -19,6 +19,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\TT.Abp.Shops\TT.Abp.Shops.csproj" />
-      <ProjectReference Include="..\TT.Core\TT.Core.csproj" />
+      <ProjectReference Include="..\TT.Extensions\TT.Extensions.csproj" />
     </ItemGroup>
 </Project>
