+++ backEnd/modules/TT.Abp.Core/TT.Abp.Core.csproj
@@ -21,7 +21,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\TT.SoMall.Domain\TT.SoMall.Domain.csproj" />
-      <ProjectReference Include="..\TT.Core\TT.Core.csproj" />
+      <ProjectReference Include="..\TT.Extensions\TT.Extensions.csproj" />
     </ItemGroup>
 
 </Project>
