+++ backEnd/modules/TT.Abp.Core/TtCoreModule.cs
@@ -5,7 +5,7 @@ using MediatR.Pipeline;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using TT.Abp.Core.Services;
-using TT.Redis;
+using TT.Extensions.Redis;
 using Volo.Abp.AspNetCore.Mvc;
 using Volo.Abp.Auditing;
 using Volo.Abp.AuditLogging.EntityFrameworkCore;
