+++ backEnd/modules/TT.Abp.VisitorManagement/TT.Abp.VisitorManagement.csproj
@@ -17,7 +17,7 @@
       <ProjectReference Include="..\..\modules\TT.HttpClient.Weixin\TT.HttpClient.Weixin.csproj" />
       <ProjectReference Include="..\TT.Abp.AppManagement\TT.Abp.AppManagement.csproj" />
       <ProjectReference Include="..\TT.Abp.Shops\TT.Abp.Shops.csproj" />
-      <ProjectReference Include="..\TT.Core\TT.Core.csproj" />
+      <ProjectReference Include="..\TT.Extensions\TT.Extensions.csproj" />
     </ItemGroup>
 
 </Project>
