+++ backEnd/modules/TT.Abp.Weixin/TT/Abp/Weixin/Domain/WeixinManager.cs
@@ -8,11 +8,11 @@ using TT.Abp.OssManagement;
 using TT.Abp.Shops.Application;
 using TT.Abp.Weixin.Application;
 using TT.Extensions;
+using TT.Extensions.Oss;
+using TT.Extensions.Redis;
 using TT.HttpClient.Weixin;
 using TT.HttpClient.Weixin.Helpers;
 using TT.HttpClient.Weixin.WeixiinResult;
-using TT.Oss;
-using TT.Redis;
 using Volo.Abp;
 using Volo.Abp.Caching;
 using Volo.Abp.DependencyInjection;
