--- backEnd/modules/TT.Core/TT.Core.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <Import Project="..\..\common.props"/>
-    <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <RootNamespace>TT</RootNamespace>
-        <OutputType>Library</OutputType>
-        <Authors>tt-tjw@qq.com</Authors>
-        <VersionPrefix>1.0.0</VersionPrefix>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.4"/>
-        <PackageReference Include="Newtonsoft.Json" Version="12.0.3"/>
-        <PackageReference Include="StackExchange.Redis" Version="2.1.30"/>
-    </ItemGroup>
-</Project>
-
-
-
