+++ backEnd/modules/TT.Extensions/RandomExt.cs
@@ -1,0 +1,60 @@
+using System;
+
+namespace TT.Extensions
+{
+    /// <summary>
+    /// A shortcut to use <see cref="Random"/> class.
+    /// Also provides some useful methods.
+    /// </summary>
+    public static class RandomExt
+    {
+        private static readonly Random Rnd = new Random();
+
+        /// <summary>
+        /// Returns a random number within a specified range.
+        /// </summary>
+        /// <param name="minValue">The inclusive lower bound of the random number returned.</param>
+        /// <param name="maxValue">The exclusive upper bound of the random number returned. maxValue must be greater than or equal to minValue.</param>
+        /// <returns>
+        /// A 32-bit signed integer greater than or equal to minValue and less than maxValue; 
+        /// that is, the range of return values includes minValue but not maxValue. 
+        /// If minValue equals maxValue, minValue is returned.
+        /// </returns>
+        public static int GetRandom(int minValue, int maxValue)
+        {
+            lock (Rnd)
+            {
+                return Rnd.Next(minValue, maxValue);
+            }
+        }
+
+        /// <summary>
+        /// Returns a nonnegative random number less than the specified maximum.
+        /// </summary>
+        /// <param name="maxValue">The exclusive upper bound of the random number to be generated. maxValue must be greater than or equal to zero.</param>
+        /// <returns>
+        /// A 32-bit signed integer greater than or equal to zero, and less than maxValue; 
+        /// that is, the range of return values ordinarily includes zero but not maxValue. 
+        /// However, if maxValue equals zero, maxValue is returned.
+        /// </returns>
+        public static int GetRandom(int maxValue)
+        {
+            lock (Rnd)
+            {
+                return Rnd.Next(maxValue);
+            }
+        }
+
+        /// <summary>
+        /// Returns a nonnegative random number.
+        /// </summary>
+        /// <returns>A 32-bit signed integer greater than or equal to zero and less than <see cref="int.MaxValue"/>.</returns>
+        public static int GetRandom()
+        {
+            lock (Rnd)
+            {
+                return Rnd.Next();
+            }
+        }
+    }
+}
