+++ backEnd/modules/TT.Extensions/TT.Extensions.csproj
@@ -1,20 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
     <PropertyGroup>
         <TargetFramework>netcoreapp3.1</TargetFramework>
-        <AssemblyName>TT.HttpClient</AssemblyName>
+        <OutputType>Library</OutputType>
+        <RootNamespace>TT.Extensions</RootNamespace>
     </PropertyGroup>
-
     <ItemGroup>
-        <PackageReference Include="AutoMapper" Version="9.0.0" />
         <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.4" />
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-        <PackageReference Include="Serilog" Version="2.9.0" />
         <PackageReference Include="StackExchange.Redis" Version="2.1.30" />
     </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\TT.Extensions\TT.Extensions.csproj" />
-    </ItemGroup>
-
 </Project>
+
+
+
