+++ backEnd/src/TT.SoMall.HttpApi.Host/Controllers/HomeController.cs
@@ -7,8 +7,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using TT.Abp.OssManagement;
-using TT.Oss;
-using TT.Redis;
+using TT.Extensions.Oss;
+using TT.Extensions.Redis;
 using Volo.Abp.AspNetCore.Mvc;
 using Volo.Abp.Auditing;
 using Volo.Abp.Identity;
