+++ backEnd/src/TT.SoMall.HttpApi/TT.SoMall.HttpApi.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\modules\TT.Core\TT.Core.csproj" />
+    <ProjectReference Include="..\..\modules\TT.Extensions\TT.Extensions.csproj" />
     <ProjectReference Include="..\TT.SoMall.Application.Contracts\TT.SoMall.Application.Contracts.csproj" />
   </ItemGroup>
